Projects/blog/newblog/requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
cffi==1.11.2
cryptography==2.1.3
Django==1.10.6
django-haystack==2.6.1
Fabric3==1.13.1.post1
gunicorn==19.7.1
idna==2.6
jieba==0.39
Markdown==2.6.9
paramiko==2.3.1
pyasn1==0.3.7
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.0
pytz==2017.3
six==1.11.0
Whoosh==2.7.4
